--- test/integration/mortgage-agent/mortgage_agent_playground_test.go
@@ -1,29 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package integration
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestAssertPlaygroundTurns(t *testing.T) {
-	a := assert.New(t)
-	assertPlaygroundTurns(t, a, playgroundTurns{
-		Turn1: `{"function_call": {"name": "legacy_dms_search_documents"}} The Sterling household 2024 tax return shows AGI of 186700. Income meets 2026 DTI. SSN: [US_SOCIAL_SECURITY_NUMBER].`,
-		Turn2: `{"function_call": {"name": "corporate_email_send_email"}} I sent the summary email to the requested address. status=sent message_id=abc`,
-	})
-}
